+++ changelog.md
@@ -12,6 +12,7 @@
 - Apply Damage fails when any token in the scene has no actor #2169
 - Some skills not recognized if not quoted in OtF #2163
 - On import from GCS, default skill cause unspent points, and skill points to be empty #1674
+- Dragging a damage value in chat to a token bigger than 1x1 on the map does not open the damage calculator #2128
 
 ## Release 0.17.18 02/09/2025